Refactor Dockerfile
This commit fixes an issue with the Dockerfile previously where the jar
needed to be built on the local machine and copied into the docker
container. This causes issues if the local machine is unable to build the
correct jar and is not in line with containerization principles.

Additionally, fix the compose.yaml to properly expose the correct ports so
that the main class does not need to be run again to correctly expose and
forward the ports.

Lastly add, an env sample file so that the correct envs can be passed into
compose.yaml so that the correct env variables make it into the docker
container to be used by the jar.