Add UserMetadata API
This commit adds the user metadata API. This API contains two REST
methods: GET and POST for /userMetadata. POST is an upsert that adds or
replaces a given user's metadata entry in the mongodb collection using the
user's google ID as a key. GET retrieves the user's metadata given a
user's ID.