Stop accepting author ID for review and fix top n reviews endpoint
This commit adds some fixes:
1. Stops accepting author ID for a new review which in turn stops writing
it to the database. This is because this field is deprecated with the new
username changes, as usernames are stored in a separate table now.
2. Top N reviews endpoint does not perform joins on the user metadata
table to retrieve username information. New code has been added to fix
that and refactors have been done to make this process easier for future
endpoints.