Refactor userMetadata instances in code to reflect public status
This commit refactors mention of userMetadata in the code to instead
refer to such data as public metadata. This distinction is important to
get right now while we still have a small userbase and the code is
relatively simple. This is to prevent a larger refactor in the future
when we have user data that only its own user can access.